requirements.txt
pandas>=2.0.0
pydantic>=2.0.0
pyyaml>=6.0
rapidfuzz>=3.0.0
typer>=0.9.0
rich>=13.0.0

